reorganize experiment + logger hierarchy, exception message changes
